// lib/api-zod/src/generated/types/appointmentRequestStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalOrderHub - Multi-tenant local marketplace API
 * OpenAPI spec version: 0.1.0
 */

export type AppointmentRequestStatus = typeof AppointmentRequestStatus[keyof typeof AppointmentRequestStatus];


export const AppointmentRequestStatus = {
  NEW: 'NEW',
  CONFIRMED: 'CONFIRMED',
  DECLINED: 'DECLINED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED',
} as const;